# src/solutions/CopilotAgent/spaarke-bff-openapi.yaml
openapi: 3.0.3
info:
  title: Spaarke BFF API
  description: >
    Spaarke Backend-for-Frontend API for legal document management, AI-powered analysis,
    matter tracking, and playbook execution. Used by M365 Copilot Declarative Agent
    to provide Spaarke AI capabilities through the Copilot side pane.
  version: 1.0.0
servers:
  - url: https://spaarke-bff-dev.azurewebsites.net
    description: Development environment
paths:
...
  /api/v1/documents:
    get:
      operationId: listDocuments
      summary: List documents
      description: Returns documents accessible to the current user. Supports filtering by matter, project, and document type.
      parameters:
        - name: matterId
          in: query
          schema:
            type: string
            format: uuid
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: documentType
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
  /api/v1/documents/{id}:
    get:
      operationId: getDocument
      summary: Get document details
      description: Returns metadata and properties for a specific document.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
...
  /api/communications/send:
    post:
      operationId: sendCommunication
      summary: Send a communication
      description: Sends an email or other communication. Use when a user asks to draft and send an email about a matter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCommunicationRequest'
      responses:
        '200':
          description: Communication sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunicationResult'

  # ─── Finance ───
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        documentType:
          type: string
        matterName:
          type: string
        matterId:
          type: string
          format: uuid
        uploadedAt:
          type: string
          format: date-time
        fileSize:
          type: integer
        pageCount:
          type: integer
...
    SendCommunicationRequest:
      type: object
      properties:
        recipientEmail:
          type: string
        subject:
          type: string
        body:
          type: string
        matterId:
          type: string
          format: uuid
    CommunicationResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string

    # ─── Finance ───
...
  securitySchemes:
    oAuth2AuthCode:
      type: oauth2
      description: OAuth configuration for the Spaarke BFF API
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/a221a95e-6abc-4434-aecc-e48338a1b2f2/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/a221a95e-6abc-4434-aecc-e48338a1b2f2/oauth2/v2.0/token
          scopes:
            api://1e40baad-e065-4aea-a8d4-4b7ab273458c/access_as_user: Access Spaarke BFF API as the user
security:
  - oAuth2AuthCode:
      - api://1e40baad-e065-4aea-a8d4-4b7ab273458c/access_as_user